requirements.txt
aiosqlite~=0.17.0
alembic~=1.7.7
discord.py~=2.1.0
python-dotenv~=0.13.0
sqlalchemy[asyncio]~=1.4.32
